pyproject.toml
[project]
name = "notes-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "openai>=2.8.1",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.12",
    "pypdf>=6.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]
